+++ tests/maud.rs
@@ -128,16 +128,14 @@ mod elements_and_attributes {
     }
 
     #[test]
-    #[ignore]
     fn custom_elements_and_data_attributes() {
-        unimplemented!()
         // assert_tmpl!({
         //     article data-index="12345" {
         //         h1 { "My blog" }
         //         tag-cloud { "pinkie pie pony cute" }
         //     }
         // }, {
-        //     article[data-index="12345"] {
+        //     article["data-index"="12345"] {
         //         h1[] { "My blog" }
         //         tag-cloud[] { "pinkie pie pony cute" }
         //     }